requirements.txt
streamlit
supabase
streamlit-url-fragment
pyjwt
streamlit-js